requirements.txt
requests>=2.28.0
pyjwt>=2.6.0
cryptography>=38.0.0
pyyaml>=6.0
urllib3>=1.26.0
